+++ Cargo.toml
@@ -32,6 +32,7 @@ solders-macros = { workspace = true }
 solders-primitives = { workspace = true }
 solders-account-decoder = { workspace = true }
 solders-rpc-filter = { workspace = true }
+solders-rpc-config = { workspace = true }
 solders-transaction-status = { workspace = true }
 solana-program-test = { workspace = true, optional = true }
 solana-banks-client = { workspace = true, optional = true }
@@ -58,7 +59,8 @@ members = [
     "primitives",
     "account-decoder",
     "transaction-status",
-    "rpc-filter"
+    "rpc-filter",
+    "rpc-config"
 ]
 
 [workspace.dependencies]
@@ -68,6 +70,7 @@ solders-traits = { path = "./traits", version = "0.14.4" }
 solders-account-decoder = { path = "./account-decoder", version = "0.14.4" }
 solders-transaction-status = { path = "./transaction-status", version = "0.14.4" }
 solders-rpc-filter = { path = "./rpc-filter", version = "0.14.4" }
+solders-rpc-config = { path = "./rpc-config", version = "0.14.4" }
 serde_with = { version = "^1.14.0", features = ["base64"] }
 solders-primitives = { path = "./primitives", version = "0.14.4" }
 solana-sdk = "^1.15.2"
