+++ rpc-config/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "solders-rpc-config"
+version = "0.14.4"
+edition = "2021"
+include = ["/src"]
+license = "MIT"
+repository = "https://github.com/kevinheavey/solders"
+
+[dependencies]
+pyo3 = { workspace = true }
+serde = { workspace = true }
+solders-macros = { workspace = true }
+solders-traits = { workspace = true }
+solders-account-decoder = { workspace = true }
+solders-transaction-status = { workspace = true }
+solders-rpc-filter = { workspace = true }
+solders-primitives = { workspace = true }
+solana-sdk = { workspace = true }
